\name{calcWhichCell}
\alias{calcWhichCell}
\title{calculating cell number}
\usage{
calcWhichCell(rc, whichRow = NULL, whichCol = NULL)
}
\arguments{
  \item{rc}{integer vector with 2 elements giving the
  dimension of the ecological inference table}

  \item{whichRow}{integer naming the specified row number}

  \item{whichCol}{integer naming the specified column
  number}
}
\value{
returns vector of innerCells
}
\description{
Calculates Cell numbers for whichCell-parameters in
eiwild-plot function
}

